Add files via upload
This app was made with Python, this was a personal project during my free time in my university years. It is about a character who must throw an object towards a vehicle, which appears in random positions. the user must indicate the initial conditions, for example, initial velocity and acceleration, both in the X and Y axes.